+++ frontend/src/Components/Board/Board.jsx
@@ -2,108 +2,13 @@ import React, {useEffect, useState} from 'react';
 import style from './Board.module.css'
 import {Link, useParams} from "react-router-dom";
 import axios from "axios";
-import formatDate from "../../Utils/helpers"
+import {formatDate, getButtonColor, getFontColor} from "../../Utils/helpers"
 import Pagination from "../Pagenation/Pagination";
+import BoardTop from "../BoardTop/BoardTop";
 
 function getUrl(subUrl) {
     const urlRoot = 'http://127.0.0.1:8000'
     return `${urlRoot}${subUrl}`
-}
-
-const mbtiParams = {
-    "intj": {
-        image: require("../../Assets/images/intj.jpg"), mainColor: "#F0DCFF",
-        fontColor: "#DDB0FF", buttonColor: "#DDB0FF"
-    },
-    "intp": {
-        image: require("../../Assets/images/intp.jpg"), mainColor: "#F0DCFF",
-        fontColor: "#DDB0FF", buttonColor: "#DDB0FF"
-    },
-    "entj": {
-        image: require("../../Assets/images/entj.jpg"), mainColor: "#F0DCFF",
-        fontColor: "#DDB0FF", buttonColor: "#DDB0FF"
-    },
-    "entp": {
-        image: require("../../Assets/images/entp.jpg"), mainColor: "#F0DCFF",
-        fontColor: "#DDB0FF", buttonColor: "#DDB0FF"
-    },
-    "esfp": {
-        image: require("../../Assets/images/esfp.jpg"), mainColor: "#FFF3DC",
-        fontColor: "#FFA800", buttonColor: "#FFA800"
-    },
-    "estp": {
-        image: require("../../Assets/images/estp.jpg"), mainColor: "#FFF3DC",
-        fontColor: "#FFA800", buttonColor: "#FFA800"
-    },
-    "isfp": {
-        image: require("../../Assets/images/isfp.jpg"), mainColor: "#FFF3DC",
-        fontColor: "#FFA800", buttonColor: "#FFA800"
-    },
-    "istp": {
-        image: require("../../Assets/images/istp.jpg"), mainColor: "#FFF3DC",
-        fontColor: "#FFA800", buttonColor: "#FFA800"
-    },
-    "esfj": {
-        image: require("../../Assets/images/esfj.jpg"), mainColor: "#72C9CB77",
-        fontColor: "#72C9CB", buttonColor: "#72C9CB8F"
-    },
-    "isfj": {
-        image: require("../../Assets/images/isfj.jpg"), mainColor: "#72C9CB77",
-        fontColor: "#72C9CB", buttonColor: "#72C9CB8F"
-    },
-    "istj": {
-        image: require("../../Assets/images/istj.jpg"), mainColor: "#72C9CB77",
-        fontColor: "#72C9CB", buttonColor: "#72C9CB8F"
-    },
-    "estj": {
-        image: require("../../Assets/images/estj.jpg"), mainColor: "#72C9CB77",
-        fontColor: "#72C9CB", buttonColor: "#72C9CB8F"
-    },
-    "infj": {
-        image: require("../../Assets/images/infj.jpg"), mainColor: "#E1FFDC",
-        fontColor: "#73C964", buttonColor: "#73C964"
-    },
-    "infp": {
-        image: require("../../Assets/images/infp.jpg"), mainColor: "#E1FFDC",
-        fontColor: "#73C964", buttonColor: "#73C964"
-    },
-    "enfj": {
-        image: require("../../Assets/images/enfj.jpg"), mainColor: "#E1FFDC",
-        fontColor: "#73C964", buttonColor: "#73C964"
-    },
-    "enfp": {
-        image: require("../../Assets/images/enfp.jpg"), mainColor: "#E1FFDC",
-        fontColor: "#73C964", buttonColor: "#73C964"
-    },
-}
-
-const getImage = (mbti) => {
-    return mbtiParams[mbti.toLowerCase()].image
-}
-
-const getMainColor = (mbti) => {
-    return mbtiParams[mbti.toLowerCase()].mainColor
-}
-const getFontColor = (mbti) => {
-    return mbtiParams[mbti.toLowerCase()].fontColor
-}
-
-const getButtonColor = (mbti) => {
-    return mbtiParams[mbti.toLowerCase()].buttonColor
-}
-
-const BoardTop = ({mbti, description}) => {
-    return (
-        <div className={style.board_top} style={{background: getMainColor(mbti)}}>
-            <div className={style.board_top_text_container}>
-                <h3>{mbti}</h3>
-                <p>{description}</p>
-            </div>
-            <div className={style.board_top_image}>
-                <img className={style.board_top_image_size} src={getImage(mbti)} alt=""/>
-            </div>
-        </div>
-    )
 }
 
 const BoardPost = ({post, mbti}) => {
@@ -115,7 +20,7 @@ const BoardPost = ({post, mbti}) => {
                         <p style={{color: getFontColor(mbti)}}>{post.category}</p>
                     </div>
                     <div className={style.board_post_title}>
-                        <Link to={`/detail/${post.id}`}>{post.title}</Link>
+                        <Link to={`/detail/${post.id}`} state={{mbti: mbti}}>{post.title}</Link>
                         <p style={{color: getFontColor(mbti)}}>[{post.hits}]</p>
                     </div>
                     <div className={style.board_post_bottom}>
@@ -195,15 +100,15 @@ const BoardCategory = ({filter, order, onCategoryChanged}) => {
         <div>
             <div className={style.board_category}>
                 <div className={style.board_category_sub}>
-                    <button style={ categoryButtonFontStyle('')} onClick={() => handleFilter('')}>전체글</button>
-                    <button style={ categoryButtonFontStyle('질문')} onClick={() => handleFilter('질문')}>질문있어요</button>
-                    <button style={ categoryButtonFontStyle('유머')} onClick={() => handleFilter('유머')}>유머</button>
-                    <button style={ categoryButtonFontStyle('창작')} onClick={() => handleFilter('창작')}>창작</button>
+                    <button style={categoryButtonFontStyle('')} onClick={() => handleFilter('')}>전체글</button>
+                    <button style={categoryButtonFontStyle('질문')} onClick={() => handleFilter('질문')}>질문있어요</button>
+                    <button style={categoryButtonFontStyle('유머')} onClick={() => handleFilter('유머')}>유머</button>
+                    <button style={categoryButtonFontStyle('창작')} onClick={() => handleFilter('창작')}>창작</button>
                 </div>
                 <div className={style.board_category_sub}>
-                    <button style={ orderButtonFontStyle('recent')} onClick={() => handleOrder('recent')}>최신순</button>
-                    <button style={ orderButtonFontStyle('like')} onClick={() => handleOrder('like')}>추천순</button>
-                    <button style={ orderButtonFontStyle('comment')} onClick={() => handleOrder('comment')}>댓글순</button>
+                    <button style={orderButtonFontStyle('recent')} onClick={() => handleOrder('recent')}>최신순</button>
+                    <button style={orderButtonFontStyle('like')} onClick={() => handleOrder('like')}>추천순</button>
+                    <button style={orderButtonFontStyle('comment')} onClick={() => handleOrder('comment')}>댓글순</button>
                 </div>
             </div>
             <hr className={style.thick_line}/>
