+++ frontend/src/Components/BoardTop/BoardTop.module.css
@@ -1,0 +1,34 @@
+.board_top {
+    display: flex;
+    align-items: center;
+    margin: 10px 0;
+    /*height: 200px;*/
+    padding: 25px 40px;
+    width: 1200px;
+    /*background-color: #E1FFDC;*/
+}
+
+.board_top_text_container {
+    display: flex;
+    flex-direction: column;
+}
+
+.board_top_text_container > h3 {
+    font-size: 50px;
+    font-weight: bold;
+    margin-bottom: 10px;
+}
+
+.board_top_text_container > p {
+    font-size: 30px;
+    color: #969696;
+}
+
+.board_top_image {
+    margin-left: 30px;
+}
+
+.board_top_image_size {
+    width: 150px;
+    height: 150px;
+}
